refactor: Split Action back out
