+++ crates/fs_snapshot/src/lib.rs
@@ -1,3 +1,4 @@
+mod action;
 mod data;
 mod error;
 
@@ -7,6 +8,7 @@ pub mod utils;
 #[cfg(feature = "harness")]
 pub mod harness;
 
+pub use action::Action;
 pub use data::Data;
 pub use error::Error;
